Move download to repository actions
